+++ src/pg/sql/10_moran.sql
@@ -17,7 +17,7 @@ AS $$
                            num_ngbrs, permutations, geom_col, id_col)
 $$ LANGUAGE plpythonu VOLATILE PARALLEL UNSAFE;
 
--- Moran's I Local (internal function)
+-- Moran's I Local (internal function) - DEPRECATED
 CREATE OR REPLACE FUNCTION
   _CDB_AreasOfInterestLocal(
       subquery TEXT,
@@ -27,16 +27,82 @@ CREATE OR REPLACE FUNCTION
       permutations INT,
       geom_col TEXT,
       id_col TEXT)
-RETURNS TABLE (moran NUMERIC, quads TEXT, significance NUMERIC, rowid INT, vals NUMERIC)
+RETURNS TABLE (
+    moran NUMERIC,
+    quads TEXT,
+    significance NUMERIC,
+    rowid INT,
+    vals NUMERIC)
 AS $$
   from crankshaft.clustering import Moran
   moran = Moran()
-  # TODO: use named parameters or a dictionary
-  return moran.local_stat(subquery, column_name, w_type,
-                          num_ngbrs, permutations, geom_col, id_col)
+  result = moran.local_stat(subquery, column_name, w_type,
+                            num_ngbrs, permutations, geom_col, id_col)
+  # remove spatial lag
+  return [(r[6], r[0], r[1], r[7], r[5]) for r in result]
 $$ LANGUAGE plpythonu VOLATILE PARALLEL UNSAFE;
 
+-- Moran's I Local (internal function)
+CREATE OR REPLACE FUNCTION
+  _CDB_MoransILocal(
+      subquery TEXT,
+      column_name TEXT,
+      w_type TEXT,
+      num_ngbrs INT,
+      permutations INT,
+      geom_col TEXT,
+      id_col TEXT)
+RETURNS TABLE (
+    quads TEXT,
+    significance NUMERIC,
+    spatial_lag NUMERIC,
+    spatial_lag_std NUMERIC,
+    orig_val NUMERIC,
+    orig_val_std NUMERIC,
+    moran_stat NUMERIC,
+    rowid INT)
+AS $$
+
+from crankshaft.clustering import Moran
+moran = Moran()
+return moran.local_stat(subquery, column_name, w_type,
+                        num_ngbrs, permutations, geom_col, id_col)
+
+$$ LANGUAGE plpythonu VOLATILE PARALLEL UNSAFE;
+
+
 -- Moran's I Local (public-facing function)
+--  Replaces CDB_AreasOfInterestLocal
+CREATE OR REPLACE FUNCTION
+  CDB_MoransILocal(
+    subquery TEXT,
+    column_name TEXT,
+    w_type TEXT DEFAULT 'knn',
+    num_ngbrs INT DEFAULT 5,
+    permutations INT DEFAULT 99,
+    geom_col TEXT DEFAULT 'the_geom',
+    id_col TEXT DEFAULT 'cartodb_id')
+RETURNS TABLE (
+    quads TEXT,
+    significance NUMERIC,
+    spatial_lag NUMERIC,
+    spatial_lag_std NUMERIC,
+    orig_val NUMERIC,
+    orig_val_std NUMERIC,
+    moran_stat NUMERIC,
+    rowid INT)
+AS $$
+
+  SELECT
+    quads, significance, spatial_lag, spatial_lag_std,
+    orig_val, orig_val_std, moran_stat, rowid
+  FROM cdb_crankshaft._CDB_MoransILocal(
+    subquery, column_name, w_type,
+    num_ngbrs, permutations, geom_col, id_col);
+
+$$ LANGUAGE SQL VOLATILE PARALLEL UNSAFE;
+
+-- Moran's I Local (public-facing function) - DEPRECATED
 CREATE OR REPLACE FUNCTION
   CDB_AreasOfInterestLocal(
     subquery TEXT,
@@ -132,7 +198,7 @@ AS $$
 $$ LANGUAGE plpythonu VOLATILE PARALLEL UNSAFE;
 
 
--- Moran's I Local Rate (internal function)
+-- Moran's I Local Rate (internal function) - DEPRECATED
 CREATE OR REPLACE FUNCTION
   _CDB_AreasOfInterestLocalRate(
       subquery TEXT,
@@ -144,15 +210,22 @@ CREATE OR REPLACE FUNCTION
       geom_col TEXT,
       id_col TEXT)
 RETURNS
-TABLE(moran NUMERIC, quads TEXT, significance NUMERIC, rowid INT, vals NUMERIC)
+TABLE(
+    moran NUMERIC,
+    quads TEXT,
+    significance NUMERIC,
+    rowid INT,
+    vals NUMERIC)
 AS $$
   from crankshaft.clustering import Moran
   moran = Moran()
   # TODO: use named parameters or a dictionary
-  return moran.local_rate_stat(subquery, numerator, denominator, w_type, num_ngbrs, permutations, geom_col, id_col)
+  result = moran.local_rate_stat(subquery, numerator, denominator, w_type, num_ngbrs, permutations, geom_col, id_col)
+  # remove spatial lag
+  return [(r[6], r[0], r[1], r[7], r[4]) for r in result]
 $$ LANGUAGE plpythonu VOLATILE PARALLEL UNSAFE;
 
--- Moran's I Local Rate (public-facing function)
+-- Moran's I Local Rate (public-facing function) - DEPRECATED
 CREATE OR REPLACE FUNCTION
   CDB_AreasOfInterestLocalRate(
       subquery TEXT,
@@ -169,6 +242,75 @@ AS $$
 
   SELECT moran, quads, significance, rowid, vals
   FROM cdb_crankshaft._CDB_AreasOfInterestLocalRate(subquery, numerator, denominator, w_type, num_ngbrs, permutations, geom_col, id_col);
+
+$$ LANGUAGE SQL VOLATILE PARALLEL UNSAFE;
+
+-- Internal function
+CREATE OR REPLACE FUNCTION
+  _CDB_MoransILocalRate(
+      subquery TEXT,
+      numerator TEXT,
+      denominator TEXT,
+      w_type TEXT,
+      num_ngbrs INT,
+      permutations INT,
+      geom_col TEXT,
+      id_col TEXT)
+RETURNS
+TABLE(
+    quads TEXT,
+    significance NUMERIC,
+    spatial_lag NUMERIC,
+    spatial_lag_std NUMERIC,
+    orig_val NUMERIC,
+    orig_val_std NUMERIC,
+    moran_stat NUMERIC,
+    rowid INT)
+AS $$
+from crankshaft.clustering import Moran
+moran = Moran()
+return moran.local_rate_stat(
+    subquery,
+    numerator,
+    denominator,
+    w_type,
+    num_ngbrs,
+    permutations,
+    geom_col,
+    id_col
+)
+$$ LANGUAGE plpythonu VOLATILE PARALLEL UNSAFE;
+
+-- Moran's I Rate
+-- Replaces CDB_AreasOfInterestLocalRate
+CREATE OR REPLACE FUNCTION
+  CDB_MoransILocalRate(
+      subquery TEXT,
+      numerator TEXT,
+      denominator TEXT,
+      w_type TEXT DEFAULT 'knn',
+      num_ngbrs INT DEFAULT 5,
+      permutations INT DEFAULT 99,
+      geom_col TEXT DEFAULT 'the_geom',
+      id_col TEXT DEFAULT 'cartodb_id')
+RETURNS
+TABLE(
+    quads TEXT,
+    significance NUMERIC,
+    spatial_lag NUMERIC,
+    spatial_lag_std NUMERIC,
+    orig_val NUMERIC,
+    orig_val_std NUMERIC,
+    moran_stat NUMERIC,
+    rowid INT)
+AS $$
+
+SELECT
+  quads, significance, spatial_lag, spatial_lag_std,
+  orig_val, orig_val_std, moran_stat, rowid
+FROM cdb_crankshaft._CDB_MoransILocalRate(
+  subquery, numerator, denominator, w_type,
+  num_ngbrs, permutations, geom_col, id_col);
 
 $$ LANGUAGE SQL VOLATILE PARALLEL UNSAFE;